+++ packages/uilib/src/lib/plugins/communication-explorer/selected-filter-store-functions.ts
@@ -33,26 +33,22 @@ export function changeMessageConnectionFilterDirection(incoming: boolean, outgoi
 function addOrRemoveMessageType(list: string[], messageType: string, checked: boolean): string[] {
 	if (checked) {
 
-		if (!list.includes(messageType)) 
-		{ list.push(messageType) }
+		const containsTypeAlready = list.includes(messageType)
+		if (!containsTypeAlready) {
+			list.push(messageType)
+		}
+
 		return list
 	}
 
 	return list.filter((item) => item !== messageType)
 }
 
-export function setSelectedMessageTypes(name: string, state: boolean, oldList?: string[]) {
-	if (oldList === undefined) { oldList = [] }
-	let updatedList: string[] = []
-
-	if      (name === MessageType.GOOSe)         { updatedList = addOrRemoveMessageType(oldList, MessageType.GOOSe,         state) }
-	else if (name === MessageType.MMS)           { updatedList = addOrRemoveMessageType(oldList, MessageType.MMS,           state) }
-	else if (name === MessageType.SampledValues) { updatedList = addOrRemoveMessageType(oldList, MessageType.SampledValues, state) }
-    
+export function setSelectedMessageTypes(type: MessageType, isActived: boolean) {
 	selectedIEDNode.update((value) => {
 		return {
 			...value, 
-			selectedMessageTypes: updatedList,
+			selectedMessageTypes: addOrRemoveMessageType(value.selectedMessageTypes, type, isActived),
 		}
 	})
 }